Parallelize generate_pixel_labels

+++ generate_pixel_labels.m
@@ -2,15 +2,15 @@ function generate_pixel_labels(trainPath, pixelPath, imageSize, classifier)
     fprintf("Pixel Label Images Started Generating At: %s\n", datetime('now','TimeZone','local','Format','d-MMM-y HH:mm:ss Z'));
     colmap = load("colormap.mat");
     imdsTrain = imageDatastore(trainPath, IncludeSubfolders=true, LabelSource="foldernames");
-    imgResize = transform(imdsTrain, @(x) imresize(x, imageSize)); % resize images
-    imgIdx = transform(imgResize, @(x) rgb2ind(x, colmap.cm));
+    % imgResize = transform(imdsTrain, @(x) imresize(x, imageSize)); % resize images
+    % imgIdx = transform(imgResize, @(x) rgb2ind(x, colmap.cm));
 
     % pre create destination directories
     if exist(pixelPath, "dir") ~= 7
         mkdir(pixelPath);
     end
 
-    %recursive directory creationu
+    % recursive directory creationu
     folders = dir(strjoin([ trainPath "/**/*.*"], ''));
     folders = folders([folders(:).isdir] == 1);
     folders = folders(~ismember({folders(:).name}, {'.', '..'}));
@@ -20,27 +20,28 @@ function generate_pixel_labels(trainPath, pixelPath, imageSize, classifier)
         mkdir(path);
     end
 
-    fileIdx = 0;
-    while hasdata(imgIdx)
-        img = read(imgIdx);
-        fileIdx = fileIdx + 1;
+    files = imdsTrain.Files;
+    parfor fileIdx = 1:length(files)
+        file = string(files(fileIdx));
+        destPath = string(replace(file, trainPath, pixelPath));
+        if exist(destPath, 'file') ~= 7
+            img = imread(file, "png");
+            img = imresize(img, imageSize);
+            img = rgb2ind(img, colmap.cm);
+            [rows, cols] = size(img);
 
-        [rows, cols] = size(img);
+            dat = double(reshape(img, [], 1));
+            ret = fastICA(dat', 2);
 
-        dat = double(reshape(img, [], 1));
-        ret = fastICA(dat', 2);
+            componentMatrix = reshape(ret, rows, cols, 2);
 
-        componentMatrix = reshape(ret, rows, cols, 2);
+            classes = classifier(componentMatrix);
 
-        classes = classifier(componentMatrix);
+            newI = uint8(classes(:, :, 1));
+            newI = cat(3, newI, newI, newI) * 255;
 
-        newI = uint8(classes(:, :, 1));
-        newI = cat(3, newI, newI, newI) * 255;
-
-        origPath = imdsTrain.Files(fileIdx);
-        destPath = string(replace(origPath, trainPath, pixelPath));
-
-        imwrite(newI, destPath);
+            imwrite(newI, destPath);
+        end
     end
 
     fprintf("Pixel Label Images Finished Generating At: %s\n", datetime('now','TimeZone','local','Format','d-MMM-y HH:mm:ss Z'));
